validate max autostar channels

+++ src/constants.rs
@@ -22,7 +22,7 @@ pub const MAX_MAX_CHARS: i16 = 5_000;
 pub const MAX_MIN_CHARS: i16 = 5_000;
 
 pub const MAX_ASC_EMOJIS: usize = 3;
-pub const MAX_AUTOSTAR: i16 = 3; // todo
+pub const MAX_AUTOSTAR: i64 = 3;
 
 // Starboard Validation
 pub const MIN_REQUIRED: i16 = 1;
